requirements.txt
jsonschema==4.18.0
pandas==2.1.0
numpy==1.26.0
matplotlib==3.8.1
seaborn==0.12.3
torch==2.1.0+cpu
faiss-cpu==1.8.0
spacy==3.6.2
tqdm==4.66.0
datasets==2.14.0
sentence-transformers==2.3.0
transformers==4.32.0
langchain==0.0.170
scikit-learn==1.4.0
gensim==4.4.0
nltk==3.9.0
streamlit==1.26.0
